Change lock file and checksum to support idempotence
In the old format, the lock/log checksum included the bash PID, so that a given
lock or log file was specific to the instance of repeat.sh currently running.
That would allow you to run the same command in the background in multiple runs
of repeat.sh at the same time.

In the new format, the lock/log checksum does not include the bash PID. This
forces repeat.sh to wait for a lock to free before it can run a command more
than once at a time. This enables you to run the same repeat.sh command multiple
times without the chance of multiple of the same command being run at the
same time.

If you want to use locking/logging but still run repeat.sh multiple times
and run multiple of the same command at once, you can change the logging
directory to something like 'logs.$$' in your shell, so each instance
of repeat.sh will use a different set of locks/logs in different directories.